+++ components/MainMenu/Statistics/Statistics.tsx
